server/requirements.txt
fastapi[standard]
uvicorn
pydantic
requests
langchain-core
langchain-community
langgraph
python-dotenv
PyGithub
langchain-huggingface
huggingface_hub[hf_xet]
peft
unsloth
#pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu126